Move WhoisUpdate job to interactor
